+++ CTF/src/main/java/ca/mcgill/science/ctf/iitems/UserHeaderItem.java
@@ -17,7 +17,7 @@ import ca.allanwang.capsule.library.logging.CLog;
 import ca.allanwang.capsule.library.swiperecyclerview.SwipeRecyclerView;
 import ca.mcgill.science.ctf.R;
 import ca.mcgill.science.ctf.api.FullUser;
-import ca.mcgill.science.ctf.auth.TepidUtils;
+import ca.mcgill.science.ctf.api.TepidUtils;
 import ca.mcgill.science.ctf.fragments.base.BaseFragment;
 import io.reactivex.SingleObserver;
 import io.reactivex.disposables.Disposable;
@@ -28,13 +28,12 @@ import io.reactivex.disposables.Disposable;
 
 public class UserHeaderItem extends AbstractItem<UserHeaderItem, UserHeaderItem.ViewHolder> {
 
-    private String username, quota, lastPrintJob;
+    private String username, quota;
     private boolean colourPrinting;
 
     public UserHeaderItem(FullUser user) {
         this.username = user.getUser().getDisplayName();
         this.quota = Integer.toString(user.getQuota());
-        this.lastPrintJob = user.getPrintJobs().get(0).getName();
         this.colourPrinting = user.getUser().getColorPrinting();
     }
 
@@ -47,7 +46,7 @@ public class UserHeaderItem extends AbstractItem<UserHeaderItem, UserHeaderItem.
     }
 
     public static void inject(ItemAdapter<UserHeaderItem> adapter, BaseFragment fragment) {
-       TepidUtils.getFullUser(fragment.getShortUser(), new SingleObserver<FullUser>() {
+        TepidUtils.getFullUser(fragment.getShortUser(), new SingleObserver<FullUser>() {
             @Override
             public void onSubscribe(Disposable d) {
 
@@ -60,8 +59,8 @@ public class UserHeaderItem extends AbstractItem<UserHeaderItem, UserHeaderItem.
                 new Handler().postDelayed(() -> {
                     SwipeRecyclerView srv = fragment.getSRV();
                     if (srv.getLayoutManager().findFirstCompletelyVisibleItemPosition() == 1) //one below header
-                        srv.smoothScrollToPosition(0);
-                }, 750);
+                        srv.smoothScrollToPosition(0, (originalSpeed) -> originalSpeed * 3);
+                }, 600);
             }
 
             @Override
@@ -76,7 +75,6 @@ public class UserHeaderItem extends AbstractItem<UserHeaderItem, UserHeaderItem.
         super.bindView(viewHolder, payloads);
         viewHolder.username.setText(username);
         viewHolder.quota.setText(quota);
-        viewHolder.lastPrintJob.setText(lastPrintJob);
         viewHolder.colourPrinting.setChecked(colourPrinting);
     }
 
@@ -85,7 +83,6 @@ public class UserHeaderItem extends AbstractItem<UserHeaderItem, UserHeaderItem.
         super.unbindView(holder);
         holder.username.setText(null);
         holder.quota.setText(null);
-        holder.lastPrintJob.setText(null);
         holder.colourPrinting.setChecked(false);
     }
 
@@ -100,8 +97,6 @@ public class UserHeaderItem extends AbstractItem<UserHeaderItem, UserHeaderItem.
         AppCompatTextView username;
         @BindView(R.id.quota)
         AppCompatTextView quota;
-        @BindView(R.id.last_print_job)
-        AppCompatTextView lastPrintJob;
         @BindView(R.id.colour_printing)
         AppCompatCheckBox colourPrinting;