+++ CTF/src/main/kotlin/ca/mcgill/science/ctf/api/Data.kt
@@ -74,4 +74,4 @@ class UserBarcode(val _id: String, val code: Long, val time: Long)
 
 val dateFormat = SimpleDateFormat("yyyy-MM-dd hh:mm:ss", Locale.CANADA)
 
-class FullUser(val user: User, val quota: Int, val printJobs:List<PrintData>)
+class FullUser(val user: User, val quota: Int)