+++ CTF/src/main/java/ca/mcgill/science/ctf/api/TepidUtils.java
@@ -1,12 +1,5 @@
 package ca.mcgill.science.ctf.api;
 
-import java.util.List;
-
-import ca.mcgill.science.ctf.api.FullUser;
-import ca.mcgill.science.ctf.api.ITepid;
-import ca.mcgill.science.ctf.api.PrintData;
-import ca.mcgill.science.ctf.api.TepidApi;
-import ca.mcgill.science.ctf.api.User;
 import io.reactivex.Single;
 import io.reactivex.SingleObserver;
 import io.reactivex.android.schedulers.AndroidSchedulers;